+++ packages/hoc-input/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.0"></a>
+# [1.0.0](https://github.com/CraveFood/farmblocks/compare/@crave/farmblocks-hoc-input@0.2.4...@crave/farmblocks-hoc-input@1.0.0) (2018-01-18)
+
+
+### Bug Fixes
+
+* **margin:** remove margin bottom from the hoc input ([cb60c10](https://github.com/CraveFood/farmblocks/commit/cb60c10))
+
+
+### Code Refactoring
+
+* **exports:** use better proptypes export names instead of just "newProps" ([031b732](https://github.com/CraveFood/farmblocks/commit/031b732))
+
+
+### Features
+
+* **display name:** wrapp the base component name on the display name ([0651408](https://github.com/CraveFood/farmblocks/commit/0651408))
+* **exports:** include the added proptypes in the exports ([fbb86c3](https://github.com/CraveFood/farmblocks/commit/fbb86c3))
+
+
+### BREAKING CHANGES
+
+* **exports:** this package no longer exports a "newProps" propTypes dictionary, check the source for the new
+export name
+* **margin:** previous behaviour was that the form input hoc included a margin bottom, now the margin is
+responsability of the user
+
+
+
+
 <a name="0.2.4"></a>
 ## [0.2.4](https://github.com/CraveFood/farmblocks/compare/@crave/farmblocks-hoc-input@0.2.3...@crave/farmblocks-hoc-input@0.2.4) (2018-01-04)